+++ App/UI/PermissionTutorial/PermissionTutorialVM.swift
@@ -516,22 +516,22 @@ extension PermissionTutorialVM.Content {
       )
     }
     
-  static var howToRetriveDigitalGreenCertificate: Self {
+  static var howToGenerateDigitalGreenCertificate: Self {
     return PermissionTutorialVM.Content(
-        title: L10n.HomeView.RetriveGreenCertificate.discoverMoreTitle,
+        title: L10n.HomeView.GenerateGreenCertificate.discoverMoreTitle,
         items: [
           .spacer(.big),
-          .textualContent(L10n.HomeView.RetriveGreenCertificate.discoverMore1, isDark: false),
+          .textualContent(L10n.HomeView.GenerateGreenCertificate.discoverMore1, isDark: false),
           .spacer(.small),
-          .textualContent(L10n.HomeView.RetriveGreenCertificate.discoverMore2, isDark: false),
+          .textualContent(L10n.HomeView.GenerateGreenCertificate.discoverMore2, isDark: false),
           .spacer(.small),
-          .textualContent(L10n.HomeView.RetriveGreenCertificate.discoverMore3, isDark: false),
+          .textualContent(L10n.HomeView.GenerateGreenCertificate.discoverMore3, isDark: false),
           .spacer(.small),
-          .textualContent(L10n.HomeView.RetriveGreenCertificate.discoverMore4, isDark: false),
+          .textualContent(L10n.HomeView.GenerateGreenCertificate.discoverMore4, isDark: false),
           .spacer(.small),
-          .textualContent(L10n.HomeView.RetriveGreenCertificate.discoverMore5, isDark: false),
+          .textualContent(L10n.HomeView.GenerateGreenCertificate.discoverMore5, isDark: false),
           .spacer(.small),
-          .textualContent(L10n.HomeView.RetriveGreenCertificate.discoverMore6, isDark: false),
+          .textualContent(L10n.HomeView.GenerateGreenCertificate.discoverMore6, isDark: false),
           .spacer(.big)
           ],
           mainActionTitle: nil,
